requirements.txt
kafka-python
pymongo
pyspark==3.5.0
numpy